% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bing}
\alias{bing}
\title{bing lexicon}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6788 rows and 2 columns.}
\usage{
bing
}
\description{
dfdf
}
\keyword{datasets}
