% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livesentiment.R
\name{liveSentiment}
\alias{liveSentiment}
\title{Interactive sentiment analysis for real time twitter feeds}
\usage{
liveSentiment()
}
\description{
This function calls a shiny based interactive user interface for analyzing 
real time twitter feeds.
}
\details{
Setting up a twitter oath (with an existing twitter account) is necessary
beforehand.
}
